requirements.txt
fastapi
uvicorn
pydantic
python-multipart
Pillow
torch
torchvision
timm
gdown
